+++ src/spellbot/integrations/convoke.py
@@ -132,6 +132,8 @@ async def fetch_convoke_link(
         payload["spellbotGamePins"] = pins
     if game["bracket"] != GameBracket.NONE.value:
         payload["bracketLevel"] = f"B{game['bracket'] - 1}"
+    if game["format"] == GameFormat.PRE_CONS.value:
+        payload["bracketLevel"] = "Precon"  # Convoke uses Bracket to indicate "pre-cons"
     if key:
         payload["password"] = key
     headers = {"user-agent": f"spellbot/{__version__}"}